refactor: move reusable admin widget from aides to ui
